# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  interfaces.NetplanNameservers:
    properties:
      addresses:
        items:
          type: string
        type: array
      search:
        items:
          type: string
        type: array
    type: object
  interfaces.NetplanRoute:
    properties:
      to:
        type: string
      via:
        type: string
    type: object
  interfaces.UpdateInterfaceStateRequest:
    properties:
      state:
        enum:
        - up
        - down
        type: string
    required:
    - state
    type: object
  interfaces.UpdateNetplanRequest:
    properties:
      addresses:
        items:
          type: string
        type: array
      dhcp4:
        type: boolean
      dhcp6:
        type: boolean
      gateway4:
        type: string
      interface:
        type: string
      nameservers:
        $ref: '#/definitions/interfaces.NetplanNameservers'
      optional:
        type: boolean
      routes:
        items:
          $ref: '#/definitions/interfaces.NetplanRoute'
        type: array
    required:
    - interface
    type: object
  interfaces.UpdatePromiscModeRequest:
    properties:
      enabled:
        type: boolean
    required:
    - enabled
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@amplifysec.net
    name: API Support
  description: Network Threat Analysis Backend API for managing network interfaces
    and system services
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: NTA Backend API
  version: "1.0"
paths:
...
  /interfaces:
    get:
      consumes:
      - application/json
      description: Get a list of all available network interfaces with their details
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List all network interfaces
      tags:
      - interfaces
  /interfaces/{name}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific network interface
      parameters:
      - description: Interface name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get network interface details
      tags:
      - interfaces
  /interfaces/{name}/check-firewall-logs:
    get:
      consumes:
      - application/json
      description: Check if firewall logs are being received on an interface using
        tcpdump (port 514)
      parameters:
      - description: Interface name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Check firewall logs with tcpdump
      tags:
      - interfaces
  /interfaces/{name}/check-network-logs:
    get:
      consumes:
      - application/json
      description: Check if network logs are being received on an interface using
        tcpdump
      parameters:
      - description: Interface name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Check network logs with tcpdump
      tags:
      - interfaces
  /interfaces/{name}/promiscuous:
    put:
      consumes:
      - application/json
      description: Enable or disable promiscuous mode on an interface
      parameters:
      - description: Interface name
        in: path
        name: name
        required: true
        type: string
      - description: Promiscuous mode update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/interfaces.UpdatePromiscModeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update promiscuous mode
      tags:
      - interfaces
  /interfaces/{name}/state:
    put:
      consumes:
      - application/json
      description: Bring an interface up or down
      parameters:
      - description: Interface name
        in: path
        name: name
        required: true
        type: string
      - description: State update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/interfaces.UpdateInterfaceStateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update interface state
      tags:
      - interfaces
  /interfaces/netplan:
    get:
      consumes:
      - application/json
      description: Retrieve the current netplan configuration
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get netplan configuration
      tags:
      - interfaces
    put:
      consumes:
      - application/json
      description: Update netplan configuration for a specific interface
      parameters:
      - description: Netplan update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/interfaces.UpdateNetplanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update netplan configuration
      tags:
      - interfaces
...
schemes:
- http
- https
swagger: "2.0"